cartmate-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
grpcio==1.59.0
grpcio-tools==1.59.0
google-cloud-aiplatform==1.38.0
websockets==12.0
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0